+++ src/games/stendhal/server/entity/npc/action/IncreaseDefXPDependentOnLevelAction.java
@@ -44,8 +44,8 @@ public class IncreaseDefXPDependentOnLevelAction implements ChatAction {
 
 	@Override
 	public void fire(final Player player, final Sentence sentence, final EventRaiser raiser) {
-		final int start = Level.getXP(player.getLevel());
-		final int next = Level.getXP(player.getLevel() + 1);
+		final long start = Level.getXP(player.getLevel());
+		final long next = Level.getXP(player.getLevel() + 1);
 		int reward = (int) (((next - start) / (def_xpDiff) / 10) / 2);
 		player.setDefXP(reward + player.getDefXP());
 		player.notifyWorldAboutChanges();
