+++ src/games/stendhal/server/script/AdminMaker.java
@@ -79,7 +79,7 @@ public class AdminMaker extends ScriptImpl {
 			player.addXP(Level.getXP(level + xlevel) - Level.getXP(level));
 
 			// set the atk and def to half the level (is a good rule of thumb)
-			final int skills = ((Level.getXP(level) + xlevel) / 2);
+			final int skills = (int) ((Level.getXP(level) + xlevel) / 2);
 			player.setAtkXP(skills);
 			player.setDefXP(skills);
 			player.incAtkXP();
